move SelectTask to a separate file
