+++ dumper/app/src/test/java/com/google/edwmigration/dumper/application/dumper/connector/oracle/OracleMetadataConnectorTest.java
@@ -21,7 +21,7 @@ import static org.junit.Assert.assertEquals;
 
 import com.google.common.collect.ImmutableList;
 import com.google.edwmigration.dumper.application.dumper.ConnectorArguments;
-import com.google.edwmigration.dumper.application.dumper.connector.oracle.OracleMetadataConnector.SelectTask;
+import com.google.edwmigration.dumper.application.dumper.connector.oracle.task.SelectTask;
 import com.google.edwmigration.dumper.application.dumper.task.Task;
 import java.util.ArrayList;
 import java.util.List;
